# docs/05-api/openapi.yaml
openapi: 3.1.0
info:
  title: 'Gold B2B API'
  version: 1.0.0
  description: |-
    Inter-dealer trading, settlement and custody for melted gold.

    Conventions that apply to every endpoint are documented in
    docs/05-api/01-conventions.md and summarised here:

    * All weights are integer milligrams of FINE gold. All money is integer rial.
      There are no decimal or floating point quantities anywhere in this API.
    * Purity is an integer in ten-thousandths: market purity 995 is 9950.
    * Timestamps are ISO-8601 in UTC.
    * Every state-changing financial request requires an `Idempotency-Key` header
      carrying a UUID. Retrying with the same key replays the stored response
      rather than repeating the action.
    * High-value operations additionally require `X-Transaction-Signature`, a TOTP
      code, and that check runs BEFORE the idempotency key is recorded so a
      corrected retry is not answered from cache.
  contact:
    name: 'Gold B2B platform team'
servers:
  - url: /api/v1
    description: 'Versioned API root'
security:
  -
    bearerAuth: {}
...
paths:
...
  /disputes:
    get:
      tags:
        - disputes
      summary: 'Index — disputes'
      operationId: getDisputes
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        422:
          $ref: '#/components/responses/DomainError'
        429:
          $ref: '#/components/responses/RateLimited'
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
    post:
      tags:
        - disputes
      summary: 'Store — disputes'
      operationId: postDisputes
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        409:
          $ref: '#/components/responses/IdempotencyConflict'
        422:
          $ref: '#/components/responses/DomainError'
        429:
          $ref: '#/components/responses/RateLimited'
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
...
  '/disputes/{dispute}/accept':
    post:
      tags:
        - disputes
      summary: 'Accept — disputes'
      operationId: postDisputesDisputeAccept
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        409:
          $ref: '#/components/responses/IdempotencyConflict'
        422:
          $ref: '#/components/responses/DomainError'
        429:
          $ref: '#/components/responses/RateLimited'
      description: 'Requires a re-signed transaction: send a current TOTP code in `X-Transaction-Signature`.'
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/TransactionSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
    parameters:
      -
        name: dispute
        in: path
        required: true
        schema:
          type: string
...
  '/disputes/{dispute}/reply':
    post:
      tags:
        - disputes
      summary: 'Reply — disputes'
      operationId: postDisputesDisputeReply
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        422:
          $ref: '#/components/responses/DomainError'
        429:
          $ref: '#/components/responses/RateLimited'
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
    parameters:
      -
        name: dispute
        in: path
        required: true
        schema:
          type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: 'A UUID minted by the client when the intent is formed, not when the request is sent. Reuse the same key on every retry of the same intent.'
    TransactionSignature:
      name: X-Transaction-Signature
      in: header
      required: true
      schema:
        type: string
      description: 'Current TOTP code. Verified before the idempotency key is recorded.'
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        maximum: 200
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: 'Opaque cursor from the previous response. Cursor paging is stable against concurrent inserts; offset paging is not.'
  schemas:
    Envelope:
      type: object
      properties:
        data:
          description: 'Resource or collection'
        meta:
          $ref: '#/components/schemas/Meta'
        links:
          type: object
          nullable: true
      required:
        - data
        - meta
    Meta:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        server_time:
          type: string
          format: date-time
        count:
          type: integer
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: INSUFFICIENT_GOLD
            message:
              type: string
              description: 'Persian, safe to show a user'
            details:
              type: object
              nullable: true
            field_errors:
              type: object
              nullable: true
          required:
            - code
            - message
        meta:
          $ref: '#/components/schemas/Meta'
...
  responses:
    DomainError:
      description: 'A business rule refused the request'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthenticated:
      description: 'Missing or expired token'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 'Authenticated but not permitted, or the resource belongs to another organization'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: 'Too many requests'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    IdempotencyConflict:
      description: 'An identical request is still in flight, or the key was reused with a different body'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'